Minor adjustment to use Hardpoint Counter as number source for HUD
Number of hardpoints is already known/counted by a component on the
ship. Forming a reference to that to avoid needing to supply the max
count in the inspector or having duplicate code for the tag count
(trivially minor optimization due to not recounting per frame, but not
really significant enough to be a real reason for the adjustment).